+++ packages/drizzle-indexeddb/src/index.ts
@@ -76,7 +76,7 @@ export {
 	createMultiClientTransport,
 	// Client
 	IDBProxyClient,
-	createProxyDbCreator,
+	createProxyIDbCreator,
 	type SyncHandler,
 	// Server
 	IDBProxyServer,